# Simd v2 - start on fresh ground with TryParse / Parse performance assessment