--- a/csFastFloat/FastDoubleParser.cs
+++ b/csFastFloat/FastDoubleParser.cs
@@ -175,7 +175,10 @@ namespace csFastFloat
     /// <param name="decimal_separator">decimal separator to be used</param>
     /// <returns>bool : true is sucessfuly parsed</returns>
     public static unsafe bool TryParseDouble(char* first, char* last, out int characters_consumed, out double result, NumberStyles styles = NumberStyles.Float, char decimal_separator = '.')
-      => TryParseNumber(first, last, out characters_consumed, out result, styles, decimal_separator);
+      {
+        result = ParseNumber(first, last, out characters_consumed,  styles, decimal_separator);
+        return true;
+      }
 
     /// <summary>
     /// Try parsing a double from a UTF-8 encoded input in the given number style
@@ -201,7 +204,12 @@ namespace csFastFloat
     /// <param name="decimal_separator">decimal separator to be used</param>
     /// <returns>bool : true is sucessfuly parsed</returns>
     public static unsafe bool TryParseDouble(byte* first, byte* last, out int characters_consumed, out double result, NumberStyles styles = NumberStyles.Float, byte decimal_separator = (byte)'.')
-      => TryParseNumber(first, last, out characters_consumed, out result, styles, decimal_separator);
+      {
+        result = ParseNumber(first, last, out characters_consumed,  styles, decimal_separator);
+        return true;
+        
+      }
+        
 
 
     /// <summary>
@@ -212,8 +220,16 @@ namespace csFastFloat
     /// <param name="decimal_separator">decimal separator to be used</param>
     /// <returns>parsed double value </returns>
     public static unsafe double ParseDouble(string s, NumberStyles styles = NumberStyles.Float, char decimal_separator = '.')
-      => ParseDouble(s, out int _, styles, decimal_separator);
+     {
+        if (s == null)
+        ThrowArgumentNull();
+        static void ThrowArgumentNull() => throw new ArgumentNullException(nameof(s));
 
+      fixed (char* pStart = s)
+      {
+         return ParseDouble(pStart, pStart + (uint)s.Length, styles, decimal_separator);
+      }
+     }
     /// <summary>
     /// Parses double from a UTF-16 encoded string in the given number style counting number of characters consumed
     /// </summary>
@@ -310,12 +326,14 @@ namespace csFastFloat
     /// <returns>parsed double value </returns>
     public static unsafe double ParseDouble(char* first, char* last, out int characters_consumed, NumberStyles styles = NumberStyles.Float, char decimal_separator = '.')
     {
-      if (TryParseDouble(first, last, out characters_consumed, out double result, styles, decimal_separator))
-      {
-         return result;
-      }
-      ThrowArgumentException();
-      throw null;
+      return ParseNumber(first, last, out characters_consumed,  styles, decimal_separator);
+        
+      // if (TryParseDouble(first, last, out characters_consumed, out double result, styles, decimal_separator))
+      // {
+      //    return result;
+      // }
+      // ThrowArgumentException();
+      // throw null;
     }
 
 
@@ -345,11 +363,12 @@ namespace csFastFloat
     /// <param name="first">char pointer to the begining of the input</param>
     /// <param name="last">char pointer to the end of the input</param>
     /// <param name="characters_consumed">number of characters consumed while parsing</param>
-    /// <param name="result">out float : parsed float value </param>
-    /// <param name="styles">allowed styles for the input string</param>
+    /// <param name="expectedFormat">allowed styles for the input string</param>
     /// <param name="decimal_separator">decimal separator to be used</param>
-    /// <returns>bool : true indicates a succesful parsing</returns>
-    internal static unsafe bool TryParseNumber(char* first, char* last, out int characters_consumed, out double result, NumberStyles styles = NumberStyles.Float, char decimal_separator = '.')
+    /// <returns>double : parsed value</returns>
+
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
+    unsafe static internal double ParseNumber(char* first, char* last, out int characters_consumed, NumberStyles expectedFormat = NumberStyles.Float, char decimal_separator = '.')
     {
       var leading_spaces = 0;
       while ((first != last) && Utils.is_ascii_space(*first))
@@ -359,22 +378,24 @@ namespace csFastFloat
       }
       if (first == last)
       {
-        result = 0;
-        characters_consumed = 0;
-        return false;
+        ThrowArgumentException();
       }
-      ParsedNumberString pns = ParsedNumberString.ParseNumberString(first, last, styles);
+      ParsedNumberString pns = ParsedNumberString.ParseNumberString(first, last, expectedFormat);
       if (!pns.valid)
       {
-        return TryHandleInvalidInput(first, last, out characters_consumed, out result);
+        
+       //  FastDoubleParser.TryHandleInvalidInput(first, last, out characters_consumed, out double x);
+       //  return x;
+
+        return FastDoubleParser.HandleInvalidInput(first, last, out characters_consumed);
+
       }
       characters_consumed = pns.characters_consumed + leading_spaces;
 
       // Next is Clinger's fast path.
       if (DoubleBinaryConstants.min_exponent_fast_path <= pns.exponent && pns.exponent <= DoubleBinaryConstants.max_exponent_fast_path && pns.mantissa <= DoubleBinaryConstants.max_mantissa_fast_path && !pns.too_many_digits)
       {
-        result = FastPath(pns);
-        return true;
+        return FastPath(pns);
       }
 
       AdjustedMantissa am = ComputeFloat(pns.exponent, pns.mantissa);
@@ -388,22 +409,18 @@ namespace csFastFloat
       // If we called compute_float<binary_format<T>>(pns.exponent, pns.mantissa) and we have an invalid power (am.power2 < 0),
       // then we need to go the long way around again. This is very uncommon.
       if (am.power2 < 0) { am = ParseLongMantissa(first, last, decimal_separator); }
-
-      result = ToFloat(pns.negative, am);
-      return true;
+      return ToFloat(pns.negative, am);
     }
-
     /// <summary>
     /// Try to parse the input (UTF-8) and compute the double value
     /// </summary>
     /// <param name="first">byte pointer to the begining of the input</param>
     /// <param name="last">byte pointer to the end of the input</param>
     /// <param name="characters_consumed">number of characters consumed while parsing</param>
-    /// <param name="result">out float : parsed float value </param>
-    /// <param name="styles">allowed styles for the input string</param>
+    /// <param name="expectedFormat">allowed styles for the input string</param>
     /// <param name="decimal_separator">decimal separator to be used</param>
-    /// <returns>bool : true indicates a succesful parsing</returns>
-    internal static unsafe bool TryParseNumber(byte* first, byte* last, out int characters_consumed, out double result, NumberStyles styles = NumberStyles.Float, byte decimal_separator = (byte)'.')
+    /// <returns>double : parsed value</returns>
+ internal unsafe static  double ParseNumber (byte* first, byte* last, out int characters_consumed, NumberStyles expectedFormat = NumberStyles.Float, byte decimal_separator = (byte)'.')
     {
       while ((first != last) && Utils.is_space(*first))
       {
@@ -413,18 +430,21 @@ namespace csFastFloat
       {
         ThrowArgumentException();
       }
-      ParsedNumberString pns = ParsedNumberString.ParseNumberString(first, last, styles);
+      ParsedNumberString pns = ParsedNumberString.ParseNumberString(first, last, expectedFormat);
       if (!pns.valid)
       {
-        return TryHandleInvalidInput(first, last, out characters_consumed, out result);
+         FastDoubleParser.TryHandleInvalidInput(first, last, out characters_consumed, out double result );
+         return  result;
+
+        //return FastDoubleParser.HandleInvalidInput(first, last, out characters_consumed);
+
       }
       characters_consumed = pns.characters_consumed;
 
       // Next is Clinger's fast path.
       if (DoubleBinaryConstants.min_exponent_fast_path <= pns.exponent && pns.exponent <= DoubleBinaryConstants.max_exponent_fast_path && pns.mantissa <= DoubleBinaryConstants.max_mantissa_fast_path && !pns.too_many_digits)
       {
-        result = FastPath(pns);
-        return true;
+        return FastPath(pns);
       }
 
       AdjustedMantissa am = ComputeFloat(pns.exponent, pns.mantissa);
@@ -438,8 +458,7 @@ namespace csFastFloat
       // If we called compute_float<binary_format<T>>(pns.exponent, pns.mantissa) and we have an invalid power (am.power2 < 0),
       // then we need to go the long way around again. This is very uncommon.
       if (am.power2 < 0) { am = ParseLongMantissa(first, last, decimal_separator); }
-      result = ToFloat(pns.negative, am);
-      return true;
+      return ToFloat(pns.negative, am);
     }
 
 
@@ -703,51 +722,48 @@ namespace csFastFloat
     }
 
 
-    internal static unsafe bool TryHandleInvalidInput(char* first, char* last, out int characters_consumed, out double result)
+    unsafe static internal double HandleInvalidInput(char* first, char* last, out int characters_consumed)
     {
       if (last - first >= 3)
       {
         if (Utils.strncasecmp(first, "nan", 3))
         {
           characters_consumed = 3;
-          result = DoubleBinaryConstants.NaN; return true;
+          return DoubleBinaryConstants.NaN;
         }
         if (Utils.strncasecmp(first, "inf", 3))
         {
           if ((last - first >= 8) && Utils.strncasecmp(first, "infinity", 8))
           {
             characters_consumed = 8;
-            result = DoubleBinaryConstants.PositiveInfinity; return true;
+            return DoubleBinaryConstants.PositiveInfinity;
           }
           characters_consumed = 3;
-          result = DoubleBinaryConstants.PositiveInfinity; return true;
+          return DoubleBinaryConstants.PositiveInfinity;
         }
         if (last - first >= 4)
         {
           if (Utils.strncasecmp(first, "+nan", 4) || Utils.strncasecmp(first, "-nan", 4))
           {
             characters_consumed = 4;
-            result = DoubleBinaryConstants.NaN; return true;
+            return DoubleBinaryConstants.NaN;
           }
           if (Utils.strncasecmp(first, "+inf", 4) ||
               Utils.strncasecmp(first, "-inf", 4))
           {
-            if ((last - first >= 9) && Utils.strncasecmp(first + 1, "infinity", 8))
+            if((last - first >= 9) && Utils.strncasecmp(first + 1, "infinity", 8))
             {
               characters_consumed = 9;
-            }
-            else
-            {
+            } else {
               characters_consumed = 4;
             }
-            result = (first[0] == '-') ? DoubleBinaryConstants.NegativeInfinity : DoubleBinaryConstants.PositiveInfinity;
-            return true;
+            return (first[0] == '-') ? DoubleBinaryConstants.NegativeInfinity : DoubleBinaryConstants.PositiveInfinity;
           }
         }
       }
-      result = 0;
+      ThrowArgumentException();
       characters_consumed = 0;
-      return false;
+      return 0d;
     }
 
     internal unsafe static bool TryHandleInvalidInput(byte* first, byte* last, out int characters_consumed, out double result)