repairs #85 
this PR 
There's slight variation on double.Parse() for canada.txt that I can't really explain.  Sometime it goes a bit higher. (up to 417...)

|                              Method |           FileName |     Mean |     Error |    StdDev |      Min | MFloat/s |     MB/s |
|------------------------------------ |------------------- |---------:|----------:|----------:|---------:|---------:|---------:|
|    'FastFloat.ParseDouble() - UTF8' |    data/canada.txt | 4.712 ms | 0.0061 ms | 0.0057 ms | 4.702 ms |    23.63 |   444.09 |
| 'FastFloat.TryParseDouble() - UTF8' |    data/canada.txt | 4.681 ms | 0.0038 ms | 0.0030 ms | 4.676 ms |    23.76 |   446.51 |
|             FastFloat.ParseDouble() |    data/canada.txt | 5.218 ms | 0.0163 ms | 0.0153 ms | 5.187 ms |    21.42 |   402.55 |
|          FastFloat.TryParseDouble() |    data/canada.txt | 5.004 ms | 0.0107 ms | 0.0089 ms | 4.989 ms |    22.27 |   418.50 |
|    'FastFloat.ParseDouble() - UTF8' |      data/mesh.txt | 1.831 ms | 0.0004 ms | 0.0003 ms | 1.830 ms |    39.89 |   338.74 |
| 'FastFloat.TryParseDouble() - UTF8' |      data/mesh.txt | 1.829 ms | 0.0008 ms | 0.0007 ms | 1.828 ms |    39.95 |   339.24 |
|             FastFloat.ParseDouble() |      data/mesh.txt | 2.053 ms | 0.0061 ms | 0.0054 ms | 2.044 ms |    35.72 |   303.32 |
|          FastFloat.TryParseDouble() |      data/mesh.txt | 1.929 ms | 0.0033 ms | 0.0029 ms | 1.926 ms |    37.91 |   321.91 |
|    'FastFloat.ParseDouble() - UTF8' | data/synthetic.txt | 5.122 ms | 0.0185 ms | 0.0173 ms | 5.096 ms |    29.44 |   553.80 |
| 'FastFloat.TryParseDouble() - UTF8' | data/synthetic.txt | 5.233 ms | 0.0056 ms | 0.0044 ms | 5.228 ms |    28.69 |   539.84 |
|             FastFloat.ParseDouble() | data/synthetic.txt | 6.611 ms | 0.0109 ms | 0.0091 ms | 6.595 ms |    22.75 |   427.92 |
|          FastFloat.TryParseDouble() | data/synthetic.txt | 6.292 ms | 0.0076 ms | 0.0068 ms | 6.282 ms |    23.88 |   449.21 |