@CarlVerret 2021-10-15
looks like a lot of change, explained by the fact that I had to refactor a bit parse / tryparse functions... and then I revert it to fastfloat .  Will groom code in a separate PR.