@CarlVerret 2021-10-15
> Do I understand correctly that you resolved some of your performance issues by manually inlining code?

Actually yes among two other things : 

1. I removed some method chaining (not sure if I should remove all of them, I'd like to get advice on that)  Been having a real hard time to determine if method chaining really does a big difference for performance.

2. I move the TryHandleInvalidInput out of the TryParseNumber back to the ParseDouble / TryParseDouble functions.  I call it as a fallback when TryParseNumber returns false.  I also detected some parts of the code where we were throwing exception even if we weren't supposed to (in try pattern).  Maybe a remaining of the old way we were parsing