@CarlVerret 2021-10-16
Quite interesting fact : I was wondering why my hand crafted benchmark was giving inconsistent results.  After looking closely at the stopwatch, I discoverd the Elapsed.Milliseconds I was measuring upon does not represent what I thought.  It only represents whole millisecond in the current second instead of the total ms ellapsed. (https://stackoverflow.com/questions/8894425/).  Now both benchmark methods give (while not exactly similar) consistent results.