@lemire 2021-10-16
Good catch.